Gamepad input capture is now a service provided to Indigo from Tyrian
